Refactored style and input element.
I'm not 100% happy with the way things came out, and I'm definitely
going to go back at some point and re-factor things to be cleaner. If
this were a professional project I would probably be putting a lot more
effort into making sure everything is right on the first go, but this is
just a fun little side project and if I don't make some forward progress
soon I will lose all motivation to keep going.