Refactored Tag class to use TypeDecorator pattern.
Originally I intended to conceal the underlying `Tag` type within the
`Entry` class and only interact with it via strings, however this turned
out to have several frustrating limitations. By using a TypeDecorator
the binary representation in the database is transparently converted
into a list of tag IDs which are then transformed into a special wrapper
class when `Entry.tags` is accessed.